requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
python-dotenv==1.0.1
python-http-client==3.3.7
requests==2.32.3
sendgrid==6.11.0
starkbank-ecdsa==2.2.0
urllib3==2.2.2
Werkzeug==3.0.3
